% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5mat.R
\name{h5mat}
\alias{h5mat}
\title{create a file connection to HDF5 matrix}
\usage{
h5mat(file, dsname = "assay001")
}
\arguments{
\item{file}{a pathname to an HDF5 file}

\item{dsname}{internal name of HDF5 matrix to use}
}
\description{
create a file connection to HDF5 matrix
}
\examples{
fn = system.file("ban_6_17/assays.h5", package="BiocSklearn")
h5mat(fn)
}
