% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{onLoad}
\alias{onLoad}
\alias{.onLoad}
\alias{SklearnEls}
\title{onLoad}
\usage{
.onLoad(libname, pkgname)
}
\description{
onLoad
}
\note{
Package installs in .GlobalEnv a list with references to numpy, pandas, h5py, and sklearn.decomposition.
}
