% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skPCA.R
\docType{class}
\name{SkDecomp-class}
\alias{SkDecomp-class}
\title{container for sklearn objects and transforms}
\description{
container for sklearn objects and transforms
}
\section{Slots}{

\describe{
\item{\code{transform}}{stored as R matrix}

\item{\code{method}}{string identifying method}

\item{\code{object}}{reference to the python object with decomposition components}
}}

