% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skPCA.R
\name{SkDecomp}
\alias{SkDecomp}
\title{constructor for SkDecomp}
\usage{
SkDecomp(transform, method)
}
\arguments{
\item{transform}{typically a numerical matrix representing a projection of the input}

\item{method}{character(1) arbitrary tag describing the decomposition}
}
\description{
constructor for SkDecomp
}
