% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skIPCA.R
\name{skIncrPCA}
\alias{skIncrPCA}
\title{use sklearn IncrementalPCA procedure}
\usage{
skIncrPCA(mat, n_components, batch_size)
}
\arguments{
\item{mat}{a matrix -- can be R matrix or numpy.ndarray}

\item{n_components}{number of PCA to retrieve}
}
\value{
matrix with rotation
}
\description{
use sklearn IncrementalPCA procedure
}
\examples{
irloc = system.file("csv/iris.csv", package="BiocSklearn")
irismat = SklearnEls$np$genfromtxt(irloc, delimiter=',')
skIncrPCA(irismat)[1:5,]
skIncrPCA(irismat, batch_size=25L)[1:5,] # slightly different
}
