% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skPartial.R
\name{skIncrPPCA}
\alias{skIncrPPCA}
\alias{skIncrPPCA,SummarizedExperiment-method}
\title{incremental partial PCA for projection of samples from SummarizedExperiment}
\usage{
skIncrPPCA(se, chunksize, n_components, assayind = 1, ...)
}
\value{
python instance of \code{sklearn.decomposition.incremental_pca.IncrementalPCA}
}
\description{
incremental partial PCA for projection of samples from SummarizedExperiment
}
\note{
Will treat samples as records and all features (rows) as attributes, projecting
to an \code{n_components}-dimensional space.  Method will acquire chunk of assay data
and transpose before computing PCA contributions.
}
\examples{
se1500 = loadHDF5SummarizedExperiment(
     system.file("hdf5/tenx_1500", package="BiocSklearn"))
lit = skIncrPPCA(se1500[, 1:50], chunksize=5, n_components=4)
round(cor(pypc <- lit$transform(dat <- t(as.matrix(assay(se1500[,1:50]))))),3)
rpc = prcomp(dat)
round(cor(rpc$x[,1:4], pypc), 3)
}
