% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SklearnEls.R
\name{SklearnEls}
\alias{SklearnEls}
\title{mediate access to python modules from sklearn.decomposition}
\usage{
SklearnEls()
}
\description{
mediate access to python modules from sklearn.decomposition
}
\note{
Returns a list with elements np (numpy), pd (pandas), h5py (h5py),
skd (sklearn.decomposition), joblib (sklearn.externals.joblib), each
referring to python modules.
}
\examples{
els = SklearnEls()
names(els$skd) # slow at first
# try py_help(els$skd$PCA) # etc.
}
