% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skPWD.R
\name{skPWD}
\alias{skPWD}
\title{use sklearn pairwise_distances procedure}
\usage{
skPWD(mat, ...)
}
\arguments{
\item{mat}{a matrix -- can be R matrix or numpy.ndarray}

\item{\dots}{additional parameters passed to sklearn.metrics.pairwise_distances, for additional information use \code{py_help() on a reticulate-imported sklearn.metrics.pairwise_distances instance.}}
}
\value{
matrix with rotation
}
\description{
use sklearn pairwise_distances procedure
}
\note{
If no additional arguments are passed, all defaults are used.
}
\examples{
irloc = system.file("csv/iris.csv", package="BiocSklearn")
data(iris)
irismat = as.matrix(iris[,1:4]) 
chk1 = skPWD(irismat)
chk1[1:4,1:5]
chk2 = skPWD(irismat, metric='manhattan')
chk2[1:4,1:5]
}
