% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5mat.R
\name{h5mat}
\alias{h5mat}
\title{create a file connection to HDF5 matrix}
\usage{
h5mat(infile, ...)
}
\arguments{
\item{infile}{a pathname to an HDF5 file}

\item{\dots}{unused}
}
\value{
instance of (S3) h5py._hl.files.File
}
\description{
create a file connection to HDF5 matrix
}
\note{
The result of this function must be used with basiliskRun with the
env argument set to bsklenv, or there is a risk of
inconsistent python modules being invoked.  This should only be used
with the persistent environment discipline of basilisk.
}
\examples{
\dontrun{
fn = system.file("ban_6_17/assays.h5", package="BiocSklearn")
h5mat(infile=fn)
}
}
