% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skPCA.R
\docType{class}
\name{SkDecomp-class}
\alias{SkDecomp-class}
\alias{getTransformed,SkDecomp-method}
\alias{getTransformed}
\title{container for sklearn objects and transforms}
\usage{
\S4method{getTransformed}{SkDecomp}(x)
}
\arguments{
\item{x}{instance of SkDecomp}
}
\value{
the getTransformed method returns a matrix
}
\description{
container for sklearn objects and transforms
}
\section{Slots}{

\describe{
\item{\code{transform}}{stored as R matrix}

\item{\code{method}}{string identifying method}
}}

\note{
In Bioc 3.11, the `object` slot is removed.  This is a consequence
of adoption of basilisk discipline for acquiring and using python resources,
which greatly increases reliability, at the expense of added complication in
handling python objects interactively in R.  We are working on restoring
this functionality but it will take time.
}
