% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skIPart.R
\name{skIncrPartialPCA}
\alias{skIncrPartialPCA}
\title{use basilisk discipline to perform partial (n_components) incremental (chunk.size) PCA with scikit.decomposition}
\usage{
skIncrPartialPCA(mat, n_components, chunk.size = 10)
}
\arguments{
\item{mat}{a matrix}

\item{n_components}{integer(1) number of PCs to compute}

\item{chunk.size}{integer(1) number of rows to use each step}
}
\description{
use basilisk discipline to perform partial (n_components) incremental (chunk.size) PCA with scikit.decomposition
}
\note{
A good source for capabilities and examples is at the \href{https://scikit-learn.org/stable/modules/decomposition.html\#decompositions}{sklearn doc site}.
}
\examples{
lk = skIncrPartialPCA(iris[,1:4], n_components=3L)
lk
head(getTransformed(lk))
}
