% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skIPCA.R
\name{skIncrPCA}
\alias{skIncrPCA}
\title{use sklearn IncrementalPCA procedure}
\usage{
skIncrPCA(mat, n_components, batch_size)
}
\arguments{
\item{mat}{a matrix -- can be R matrix or numpy.ndarray}

\item{n_components}{number of PCA to retrieve}

\item{batch_size}{number of records to use at each iteration}
}
\value{
matrix with rotation
}
\description{
use sklearn IncrementalPCA procedure
}
\examples{
irloc = system.file("csv/iris.csv", package="BiocSklearn")
irismat = SklearnEls()$np$genfromtxt(irloc, delimiter=',')
ski = skIncrPCA(irismat)
ski25 = skIncrPCA(irismat, batch_size=25L) # non-default
getTransformed(ski)[1:3,]
getTransformed(ski25)[1:3,]
}
