% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skIPCA.R
\name{skIncrPCA}
\alias{skIncrPCA}
\title{use sklearn IncrementalPCA procedure}
\usage{
skIncrPCA(mat, n_components = 2L, batch_size = 5L, ...)
}
\arguments{
\item{mat}{a matrix -- can be R matrix or numpy.ndarray, if the latter it
must be set up in a basilisk persistent environment, and that is not
currently demonstrated for this package.}

\item{n_components}{number of PCA to retrieve}

\item{batch_size}{number of records to use at each iteration}

\item{\dots}{passed to python  IncrementalPCA}
}
\value{
matrix with rotation
}
\description{
use sklearn IncrementalPCA procedure
}
\examples{
dem = skIncrPCA(iris[,1:4], batch_size=25L)
dem2 = skIncrPCA(iris[,1:4], batch_size=25L, n_components=2L)
dem
dem2
}
