% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skPCA.R
\name{skPCA}
\alias{skPCA}
\title{use sklearn PCA procedure}
\usage{
skPCA(mat)
}
\arguments{
\item{mat}{a matrix -- can be R matrix or numpy.ndarray}
}
\value{
matrix with rotation
}
\description{
use sklearn PCA procedure
}
\note{
all defaults are used
}
\examples{
irloc = system.file("csv/iris.csv", package="BiocSklearn")
irismat = SklearnEls()$np$genfromtxt(irloc, delimiter=',')
skpi = skPCA(irismat)
getTransformed(skpi)[1:5,]
}
